requirements.txt
google-genai>=0.4.0
rich>=13.0.0
requests>=2.30.0
googlesearch-python>=1.2.3
prompt_toolkit>=3.0.0
ollama>=0.4.0
beautifulsoup4>=4.12.0
beautifulsoup4>=4.12.0
googlesearch-python>=1.2.3
anthropic>=0.30.0
openai>=1.35.0
questionary>=2.0.0
fastapi>=0.110.0
uvicorn>=0.28.0
flask>=3.0.0
streamlit>=1.32.0
gradio>=4.20.0
python-dotenv>=1.0.1
